Rename export; change storage
Wallets do not persist in `localstorage`. They are saved to the object itself;
the service should decide what to do with it from there.

The name of the variable exported to Meteor is KeyRack, not Accounts. This is
to avoid conflict with the meteor accounts module. It was not named EthAccounts
to avoid conflicting with `meteor-package-accounts`.